Restore FixAtoms rows in float64 after an Alchemi relaxation
The batch position array is float32, so assigning the float64 frozen
coordinates into it truncated them: the "restored" FixAtoms rows came back
about 5e-7 A off their pinned positions at slab-sized coordinates, in x and y,
growing with the coordinate magnitude. It did not accumulate across steps
(each restore re-truncates an already-truncated value) and it flipped no
acceptance decision in the Ag(111) fixture, but every Alchemi-relaxed frozen
substrate was slightly off its constraint.

Widen to float64 before restoring, in both write-back paths (the single
structure and the batched harvest). Frozen drift goes from 5.4e-7 A to
2.0e-15 A. Found by tests/test_ag111_gcmc_golden.py on its first run, which is
also the first test ever to execute this package.

Also folds the duplicated inline FixAtoms loop into _fixed_indices.

+++ mcpy/calculators/alchemi_f_calculator.py
@@ -259,7 +259,8 @@ class AlchemiFCalculator:
         """Write energy + relaxed positions of ``rows`` (batch row indices)
         back to their originating Atoms objects, restoring FixAtoms rows."""
         e = _per_graph_energies(batch.energy, int(batch.num_graphs))
-        pos = batch.positions.detach().cpu().numpy()
+        # float64 before restoring frozen rows: see _write_back_positions.
+        pos = batch.positions.detach().cpu().numpy().astype(np.float64)
         batch_idx = batch.batch_idx.detach().cpu().numpy()
         for row in rows:
             g = int(alive[row])
